requirements.txt
pyyaml>=6.0
youtube-transcript-api>=1.0.0
